backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
cryptography==44.0.0
python-decouple==3.8
bcrypt==4.2.0
anthropic==0.42.0
